Split Telegram configuration, wire, and state types
